# Please enter the commit message for your changes. Lines starting # with '#' will be ignored, and an empty message aborts the commit. # # On branch main # Your branch is ahead of 'origin/main' by 1 commit. #   (use "git push" to publish your local commits) # # Changes to be committed: #	modified:   blockchain/contracts/Counters.sol #	modified:   blockchain/contracts/IdMedrec.sol #	modified:   blockchain/hardhat.config.js #	deleted:    blockchain/ignition/deployments/chain-1337/artifacts/IdMedRec#IDMedRec.dbg.json #	deleted:    blockchain/ignition/deployments/chain-1337/artifacts/IdMedRec#IDMedRec.json #	deleted:    blockchain/ignition/deployments/chain-1337/deployed_addresses.json #	deleted:    blockchain/ignition/deployments/chain-1337/journal.jsonl #	new file:   blockchain/ignition/deployments/chain-80002/artifacts/IdMedRec#IDMedRec.dbg.json #	new file:   blockchain/ignition/deployments/chain-80002/artifacts/IdMedRec#IDMedRec.json #	renamed:    blockchain/ignition/deployments/chain-1337/build-info/83475da6f63dd22c5cf65da72ab8542d.json -> blockchain/ignition/deployments/chain-80002/build-info/6b36466b3bd4751fcd1da8f38ad51df1.json #	new file:   blockchain/ignition/deployments/chain-80002/deployed_addresses.json #	new file:   blockchain/ignition/deployments/chain-80002/journal.jsonl #	modified:   blockchain/test/IDMedRec.js # # Changes not staged for commit: #	modified:   blockchain/contracts/IdMedrec.sol #

+++ blockchain/contracts/Counters.sol
@@ -3,15 +3,15 @@ pragma solidity ^0.8.19;
 
 library Counters {
     struct Counter {
-        uint32 _valueRecord;
-        uint32 _valuePatient;
-        uint32 _valueDoctor;
+        uint256 _valueRecord;
+        uint256 _valuePatient;
+        uint256 _valueDoctor;
     }
 
     function current(
         Counter storage counter,
         uint8 whoIs
-    ) internal view returns (uint32) {
+    ) internal view returns (uint256) {
         return
             whoIs == 1
                 ? counter._valueRecord
